The first Stopwatch
The very first version of the stopwatch. It employs jQuery
to create the two timers for Mega Health and Heavy Armor,
colloquially called 'Mega' and 'Heavy', also giving some
idea of which is which by color -- blue for the Mega and
green for the Heavy.

An assumption is made that at the beginning of a match,
both artifacts are grabbed in 5 seconds, which is why both
counters start at 35 seconds the first time. From then on,
whenever a counter hits 0, it will reset to 30.

Another way to reset a counter to 30 is to click or tap it.

Plenty of improvements are possible, but this is good for
the time being.