Blinking Warn to Color Warn
This change turns the dimmed color blinking into
a single change to a contrast-like color. This is
because blinking may be perceived by some folks
as unpleasant, and also becase it may wear off
displays it is played on. At least I mind the
lifetime of devices I use. :)

This will be given a trial and kept if fine.

The variables mega_state and heavy_state are not
needed anymore.